Add files via upload
The aim of this task is to develop an image encryption tool that manipulates pixel data to secure an image. By altering the pixel values of an image, the tool will create an encrypted version that is unintelligible without decryption. The same tool should also support the decryption process to restore the original image. This can be achieved by implementing reversible operations like pixel swapping or mathematical transformations.